+++ robotframework-ls/docs/changelog.md
@@ -3,6 +3,7 @@ New in 0.48.0 (NEXT)
 
 - Consider as defined a reference to any variable set as a global variable anywhere on the workspace. [#641](https://github.com/robocorp/robotframework-lsp/issues/641)
 - Code analysis for template arguments.
+- Arguments aren't added when applying completions for keywords in `[Template]`.
 - GitHub CoPilot inline suggestions are applied when tab is activated instead of adding 4-spaces.
 - Fixed integration with robotframework-tidy which could make formatting the same code a second time misbehave. [#687](https://github.com/robocorp/robotframework-lsp/issues/687)
 - Properly consider that `For each input work item` from `rpaframework` receives a keyword as the first parameter. [#684](https://github.com/robocorp/robotframework-lsp/issues/684)